pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "remorph-openenv"
version = "0.1.0"
description = "OpenEnv packaging for ReMorph API drift repair scenarios."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.8,<3",
  "PyYAML>=6.0,<7",
  "openenv-core[core]>=0.2.3,<0.3",
  "fastapi>=0.115,<1",
  "httpx>=0.27,<1",
  "uvicorn>=0.30,<1",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["remorph_openenv*", "server*"]
